init linux version app

+++ sms/sources/server/files.py
@@ -8,26 +8,32 @@ import threading
 from datetime import datetime
 from pathlib import Path
 from pandas.core.algorithms import mode
+from pandas.core.frame import DataFrame
 from sms.authentification import authentification
 from sms.sources.message.campagne import campagne
+from sms.color import colors
 
 class files(authentification):
     def __init__(self, accountID, token):
         super().__init__(accountID, token)
         self.directory = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
-        self.docFiles = self.directory+"\\tmp\\"
+        self.docFiles = self.directory+"//tmp//"
         
     ##Fonctionlity Util 
     def writeToLog(self, logName, data):
-        pathLog =  self.docFiles+"LOG\\"+logName+"_"+data['etat']+".txt"
+        pathLog =  self.docFiles+"LOG//"+logName+"_"+data['etat']+".txt"
         print(pathLog)
         with open(pathLog,"a") as f:
-            f.write(str(data)+";")
+            f.write(str(data)+",\n")
             
     def formDate(self,date):
         dateSplit = date.split('-')
         if dateSplit[0] != '':
-            dateShoot = dateSplit[2].split(' ')[0]+"/"+dateSplit[1]+"/"+dateSplit[0]
+            try:
+                dateShoot = dateSplit[2].split(' ')[0]+"/"+dateSplit[1]+"/"+dateSplit[0]
+            except Exception as e:
+                print("-----------error")
+                print(e)
             hour = date.split(' ')[1]
             result = dateShoot+" "+hour
         else:
@@ -51,7 +57,7 @@ class files(authentification):
     def appendContentsInFile(self, dataToAppend, fileType):
         try:
             fileName = str(datetime.now().strftime("%Y%m%d"))+"_"+fileType+"_SMS.csv"
-            newFTPfile = self.docFiles+"FTPFiles\\"+fileName
+            newFTPfile = self.docFiles+"FTPFiles//"+fileName
             dataToAppend.to_csv(newFTPfile, sep=',', mode='a', index =False, header=False)
             result = {"etat": "success", "date": str(datetime.now())}
         except Exception as e:
@@ -66,10 +72,9 @@ class files(authentification):
             if currentCamp['value']['status'] == "sent":
                 dateShoot = currentCamp['value']['scheduledAtUtc']
                 dateShoot = self.replaceCharMultiple(currentCamp['value']['scheduledAtUtc'],((":",""),(".",""),("-","")))
-                pathFolderFile = "tmp\\"+fileType+"\\"+idCampagne+"_"+dateShoot+"_"+fileType+".csv"
+                pathFolderFile = "tmp//"+fileType+"//"+idCampagne+"_"+dateShoot+"_"+fileType+".csv"
                 pathSaveFile = os.path.join(self.directory,pathFolderFile)
                 url = self.pathInitApi_V1_message+self.accountID+"/messages/"+idCampagne+"/export?status="+fileType if fileType != "optedOut" else self.pathInitApi_V2_addressbook+self.accountID+"/optout/export?campaignId="+idCampagne
-                print(url)
                 pathSaveFile = os.path.join(self.directory,pathFolderFile)
                 try:
                     confOpen = urllib.request.build_opener()
@@ -82,60 +87,69 @@ class files(authentification):
                         result = {"etat": "success", "pathFile":pathSaveFile, "idStats": idStats, "fileType": fileType}
                     else:
                         result = {"etat": "error ", "etat_description": "file not created"}
+                        print(f"{colors.FAIL} {idCampagne} ERROR Download--- in downloadFiles{colors.ENDC}")
                         
                 except Exception as e:
                     result = {"etat": "error ", "etat_description": str(e)}
+                    print(f"{colors.FAIL} {idCampagne} {str(e)}--- in downloadFiles{colors.ENDC}")
             else:
                 result = {"etat": "error ", "etat_description": "status "+ currentCamp['value']['status']}
         else:
             result = {"etat": "error ", "etat_description": "campagne not found"}
+            print(f"{colors.FAIL} {idCampagne} campagne not found--- in downloadFiles{colors.ENDC}")
             
         return result
     
     def filterFile(self, dataInfoCreatedFile):
-        fileContentFrame = pd.read_csv(dataInfoCreatedFile['pathFile'], sep=',', na_filter=False)
-        dataKonticrea = {"idStats":"04", "tag_campagne":"test", "code_pays":"fr"}
-        fileContentFrame['idStats'] = dataKonticrea['idStats']
-        fileContentFrame['tag_campagne'] = dataKonticrea['tag_campagne']
-        fileContentFrame['code_pays'] = dataKonticrea['code_pays']
-
-        if dataInfoCreatedFile['fileType'] == "optedOut":
-            data  = {
-                "date_shoot" : fileContentFrame['DateUtc'].apply(lambda date: self.formDate(str(date))),
-                "mobile" : fileContentFrame['OptOutValue'],
-                "idStats" : fileContentFrame['idStats'],
-                "tag_campagne" : fileContentFrame['tag_campagne'],
-                "code_pays" : fileContentFrame['code_pays'],
-            }
-        # elif dataInfoCreatedFile['fileType'] == "undelivered":
-        #    pass
-        else:
-            data={
-                "date_shoot" : fileContentFrame['Processed UTC'].apply(lambda date: self.formDate(str(date))),
-                "mobile" : fileContentFrame['Recipient'],
-                "idStats" : fileContentFrame['idStats'],
-                "tag_campagne" : fileContentFrame['tag_campagne'],
-                "code_pays" : fileContentFrame['code_pays'],
-            }
-            
-        resultFrame = pd.DataFrame(data)
-        # withstatus = fileContentFrame[fileContentFrame["Status"] == "Failed"]
-        # print(withstatus.at[1,"Processed UTC"])
-        # print(withstatus)
-        # print(resultFrame[resultFrame['date_shoot']==''])
-        return resultFrame
+        try:
+            fileContentFrame = pd.read_csv(dataInfoCreatedFile['pathFile'], sep=',', na_filter=False)
+            dataKonticrea = {"idStats":"04", "tag_campagne":"test", "code_pays":"fr"}
+            fileContentFrame['idStats'] = dataKonticrea['idStats']
+            fileContentFrame['tag_campagne'] = dataKonticrea['tag_campagne']
+            fileContentFrame['code_pays'] = dataKonticrea['code_pays']
+            if dataInfoCreatedFile['fileType'] == "optedOut":
+                data  = {
+                    "date_shoot" : fileContentFrame['DateUtc'].apply(lambda date: self.formDate(str(date))),
+                    "mobile" : fileContentFrame['OptOutValue'],
+                    "idStats" : fileContentFrame['idStats'],
+                    "tag_campagne" : fileContentFrame['tag_campagne'],
+                    "code_pays" : fileContentFrame['code_pays'],
+                }
+            # elif dataInfoCreatedFile['fileType'] == "undelivered":
+            #    pass
+            else:
+                data={
+                    "date_shoot" : fileContentFrame['Processed UTC'].apply(lambda date: self.formDate(str(date))),
+                    "mobile" : fileContentFrame['Recipient'],
+                    "idStats" : fileContentFrame['idStats'],
+                    "tag_campagne" : fileContentFrame['tag_campagne'],
+                    "code_pays" : fileContentFrame['code_pays'],
+                }
+                
+            resultFrame = pd.DataFrame(data)
+            # withstatus = fileContentFrame[fileContentFrame["Status"] == "Failed"]
+            # print(withstatus.at[1,"Processed UTC"])
+            # print(withstatus)
+            # print(resultFrame[resultFrame['date_shoot']==''])
+            return resultFrame
+        except Exception as e:
+            print(f"{colors.FAIL} ERROR: {str(e)}  PATH: {dataInfoCreatedFile['pathFile']} {colors.ENDC}")
+            return {'etat':'ERROR'}
     
     def launchDWHUpdate(self, campagneList,fileType):
         for idCamp in campagneList:
             infoCurrentCamp =  self.downloadFiles(idCamp,fileType)
             if "etat" in  infoCurrentCamp and infoCurrentCamp["etat"] == "success":
                 dataToAppend = self.filterFile(infoCurrentCamp)
-                appendResult = self.appendContentsInFile(dataToAppend,fileType)
-                appendResult['idCampagne'] = idCamp 
-                self.writeToLog(fileType,appendResult)
-                if appendResult['etat'] == 'success':
-                    print("remove----",str(infoCurrentCamp['pathFile']))
-                    os.remove(infoCurrentCamp['pathFile'])
+                if isinstance(dataToAppend,DataFrame):
+                    appendResult = self.appendContentsInFile(dataToAppend,fileType)
+                    appendResult['idCampagne'] = idCamp 
+                    self.writeToLog(fileType,appendResult)
+                    if appendResult['etat'] == 'success':
+                        os.remove(infoCurrentCamp['pathFile'])
+                    else:
+                        print(f"{colors.FAIL} {idCamp} error--------Fies not removed---{colors.ENDC}")
+                        print(appendResult)
     
     def executThread(self,listeCampagne):
         result = {"etat": "success"}
@@ -147,6 +161,8 @@ class files(authentification):
             theadList = [undelivered,converted,optedOut,delivered]
             [t.start() for t in theadList]
             [t.join() for t in theadList]
+            print(f"{colors.OKGREEN}--- All process executed ---{colors.ENDC}")
+            return result
         except Exception as e:
-            result = {"etat": "error ", "etat_description": str(e)}
-        return result
+            print(f"{colors.FAIL} ERROR--{str(e)} {colors.ENDC}")
+            return {"etat": "error ", "etat_description": str(e)}
